# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Xclusive API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me:
    patch:
      operationId: updateProfile
      tags: [users]
      summary: Atualizar perfil do utilizador autenticado
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Perfil atualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /users/{username}:
    get:
      operationId: getUserProfile
      tags: [users]
      summary: Obter perfil de um utilizador
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Perfil do utilizador
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "404":
          description: Utilizador não encontrado
  /users/{username}/follow:
    post:
      operationId: followUser
      tags: [users]
      summary: Seguir utilizador
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: A seguir

    delete:
      operationId: unfollowUser
      tags: [users]
      summary: Deixar de seguir utilizador
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deixou de seguir
  /users/{username}/followers:
    get:
      operationId: getUserFollowers
      tags: [users]
      summary: Seguidores do utilizador
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Lista de seguidores
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedUsers"
  /users/{username}/following:
    get:
      operationId: getUserFollowing
      tags: [users]
      summary: Utilizadores que segue
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Lista de seguidos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedUsers"
  /users/{username}/posts:
    get:
      operationId: getUserPosts
      tags: [users]
      summary: Posts de um utilizador
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Lista de posts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedPosts"
  /users/{username}/reels:
    get:
      operationId: getUserReels
      tags: [users]
      summary: Reels de um utilizador
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Lista de reels
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedReels"

  # FEED
...
components:
  schemas:
...
    UserProfile:
      type: object
      required: [id, username, nomeExibicao, tipoConta, verificado, privado, totalSeguidores, totalSeguindo, totalPublicacoes, criadoEm]
      properties:
        id:
          type: integer
        username:
          type: string
        nomeExibicao:
          type: string
        bio:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        capaUrl:
          type: ["string", "null"]
        link:
          type: ["string", "null"]
        tipoConta:
          type: string
          enum: [pessoal, criador]
        verificado:
          type: boolean
        privado:
          type: boolean
        totalSeguidores:
          type: integer
        totalSeguindo:
          type: integer
        totalPublicacoes:
          type: integer
        estaASeguir:
          type: boolean
        segueVoce:
          type: boolean
        criadoEm:
          type: string
          format: date-time
    UserSummary:
      type: object
      required: [id, username, nomeExibicao, verificado]
      properties:
        id:
          type: integer
        username:
          type: string
        nomeExibicao:
          type: string
        avatarUrl:
          type: ["string", "null"]
        verificado:
          type: boolean
        tipoConta:
          type: string
        estaASeguir:
          type: boolean
        totalSeguidores:
          type: integer
    ProfileUpdate:
      type: object
      properties:
        nomeExibicao:
          type: string
        bio:
          type: string
        link:
          type: string
        avatarUrl:
          type: string
        capaUrl:
          type: string
        tipoConta:
          type: string
          enum: [pessoal, criador]
        privado:
          type: boolean
    PaginatedUsers:
      type: object
      required: [users, total, page, hasMore]
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/UserSummary"
        total:
          type: integer
        page:
          type: integer
        hasMore:
          type: boolean

    # POSTS
    Post:
      type: object
      required: [id, autor, tipo, exclusivo, totalCurtidas, totalComentarios, guardado, curtido, criadoEm]
      properties:
        id:
          type: integer
        autor:
          $ref: "#/components/schemas/UserSummary"
        legenda:
          type: ["string", "null"]
        localizacao:
          type: ["string", "null"]
        tipo:
          type: string
          enum: [imagem, video, carrossel]
        media:
          type: array
          items:
            $ref: "#/components/schemas/PostMedia"
        exclusivo:
          type: boolean
        precoDesbloqueio:
          type: ["number", "null"]
        totalCurtidas:
          type: integer
        totalComentarios:
          type: integer
        curtido:
          type: boolean
        guardado:
          type: boolean
        criadoEm:
          type: string
          format: date-time
    PostMedia:
      type: object
      required: [id, url, tipo, ordem]
      properties:
        id:
          type: integer
        url:
          type: string
        tipo:
          type: string
          enum: [imagem, video]
        ordem:
          type: integer
...
    PaginatedPosts:
      type: object
      required: [posts, total, page, hasMore]
      properties:
        posts:
          type: array
          items:
            $ref: "#/components/schemas/Post"
        total:
          type: integer
        page:
          type: integer
        hasMore:
          type: boolean

    # COMMENTS
...
    Reel:
      type: object
      required: [id, autor, videoUrl, exclusivo, totalCurtidas, totalComentarios, curtido, guardado, criadoEm]
      properties:
        id:
          type: integer
        autor:
          $ref: "#/components/schemas/UserSummary"
        videoUrl:
          type: string
        capaUrl:
          type: ["string", "null"]
        legenda:
          type: ["string", "null"]
        somTitulo:
          type: ["string", "null"]
        somArtista:
          type: ["string", "null"]
        exclusivo:
          type: boolean
        totalCurtidas:
          type: integer
        totalComentarios:
          type: integer
        curtido:
          type: boolean
        guardado:
          type: boolean
        criadoEm:
          type: string
          format: date-time
...
    PaginatedReels:
      type: object
      required: [reels, page, hasMore]
      properties:
        reels:
          type: array
          items:
            $ref: "#/components/schemas/Reel"
        page:
          type: integer
        hasMore:
          type: boolean

    # SEARCH
...